refactor: transition of helpers on the example of helpers.core
